LayoutTests/platform/mac/accessibility/table-fallback-roles-expose-element-attributes-expected.txt

item1
item1
item1
This tests if an element is using fallback role tokens, it will still become the 'correct' element which will expose the correct attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


cell: AXRole: AXCell, AXValue:

cellChild: AXRole: AXStaticText, AXValue: item1

cell: AXRole: AXCell, AXValue:

cellChild: AXRole: AXStaticText, AXValue: item1

cell: AXRole: AXCell, AXValue:

cellChild: AXRole: AXStaticText, AXValue: item1

PASS successfullyParsed is true

TEST COMPLETE
